requirements.txt
coreapi==1.32.3
Django==1.10.1
django-rest-swagger==2.0.5
djangorestframework==3.4.6
itypes==1.1.0
openapi-codec==1.0.0
psycopg2==2.6.2
requests==2.11.1
simplejson==3.8.2
uritemplate==3.0.0
wheel==0.24.0
pillow==3.2.0
django-imagekit==3.3
factory-boy
django_filters
django-rest-auth
django-allauth